server/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
flask-cors
python-dateutil
pyjwt
flask_socketio
eventlet
